Refactored test method to PitPanelTest class
